# shared/api_spec.yaml
openapi: 3.0.3
info:
  title: DPS Dashboard API
  description: >
    Deshapriya Park Durga Puja Club Management Dashboard API.
    All endpoints require authentication unless marked as public.
    Role enforcement is server-side on every route.
  version: 1.0.0
  contact:
    name: DPS Dashboard Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: http://{vps_ip}/api
    description: LunaNode VPS
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: HTTP-only JWT session cookie set by NextAuth
  schemas:
...
    TransactionType:
      type: string
      enum: [CASH_IN, CASH_OUT]
    TransactionCategory:
      type: string
      enum: [MEMBERSHIP_FEE, APPLICATION_FEE, SPONSORSHIP, EXPENSE, OTHER]
    PaymentMode:
      type: string
      enum: [UPI, BANK_TRANSFER, CASH]
    SponsorPurpose:
      type: string
      enum: [TITLE_SPONSOR, GOLD_SPONSOR, SILVER_SPONSOR, FOOD_PARTNER, MEDIA_PARTNER, STALL_VENDOR, MARKETING_PARTNER]
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    ApprovalEntityType:
      type: string
      enum: [TRANSACTION, MEMBER_ADD, MEMBER_EDIT, MEMBER_DELETE, MEMBERSHIP]
    ApprovalSource:
      type: string
      enum: [MANUAL, RAZORPAY_WEBHOOK]

    # --- Request/Response Models ---
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
...
    TransactionUpdate:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/TransactionType'
        category:
          $ref: '#/components/schemas/TransactionCategory'
        amount:
          type: number
          format: decimal
        paymentMode:
          $ref: '#/components/schemas/PaymentMode'
        description:
          type: string
        sponsorPurpose:
          $ref: '#/components/schemas/SponsorPurpose'
        senderName:
          type: string
        senderPhone:
          type: string
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TransactionType'
        category:
          $ref: '#/components/schemas/TransactionCategory'
        amount:
          type: number
        paymentMode:
          $ref: '#/components/schemas/PaymentMode'
        description:
          type: string
        sponsorPurpose:
          $ref: '#/components/schemas/SponsorPurpose'
        memberId:
          type: string
          format: uuid
        sponsorId:
          type: string
          format: uuid
        enteredBy:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        approvalStatus:
          $ref: '#/components/schemas/ApprovalStatus'
        approvalSource:
          $ref: '#/components/schemas/ApprovalSource'
        approvedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
        approvedAt:
          type: string
          format: date-time
          nullable: true
        razorpayPaymentId:
          type: string
          nullable: true
        razorpayOrderId:
          type: string
          nullable: true
        senderName:
          type: string
          nullable: true
        senderPhone:
          type: string
          nullable: true
        senderUpiId:
          type: string
          nullable: true
        senderBankAccount:
          type: string
          nullable: true
        senderBankName:
          type: string
          nullable: true
        receiptNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    # Sponsors
...
    SponsorCheckoutResponse:
      type: object
      properties:
        token:
          type: string
        sponsorName:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        upiId:
          type: string
        bankDetails:
          type: object
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true

    # Approvals
    ApprovalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          $ref: '#/components/schemas/ApprovalEntityType'
        entityId:
          type: string
          format: uuid
        action:
          type: string
        previousData:
          type: object
          nullable: true
        newData:
          type: object
          nullable: true
        requestedBy:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        reviewedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
security:
  - cookieAuth: []
paths:
...
  /transactions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Transactions]
      summary: Get single transaction
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

    put:
      tags: [Transactions]
      summary: Update transaction
      description: >
        Admin: updates directly. Operator: creates Approval record.
        Cannot update auto-approved Razorpay transactions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: Transaction updated (or approval pending)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TransactionResponse'
                  - $ref: '#/components/schemas/ApprovalResponse'

    delete:
      tags: [Transactions]
      summary: Delete transaction
      description: >
        Admin: deletes directly. Operator: creates Approval record.
        Cannot delete auto-approved Razorpay transactions.
      responses:
        '200':
          description: Transaction deleted (or approval pending)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ==================== SPONSORS ====================
...
  /sponsor-links/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [SponsorLinks]
      summary: Get sponsor checkout data (public)
      description: "Public endpoint. Returns checkout data for the sponsor payment page."
      security: []
      responses:
        '200':
          description: Sponsor checkout data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorCheckoutResponse'
        '404':
          description: Link not found, expired, or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== APPROVALS ====================
...